first level code review incorporation and some refactoring
