+++ api/router/pubsub/ApplicationStatusHandler.go
@@ -243,12 +243,7 @@ func (impl *ApplicationStatusHandlerImpl) updateArgoAppDeleteStatus(app *v1alpha
 		}
 	} else {
 		// devtron app
-		cdPipeline, err := impl.pipelineBuilder.GetCdPipelineById(pipeline.Id)
-		if err != nil {
-			impl.logger.Errorw("error in getting cdPipeline by id", "err", err, "id", pipeline.Id)
-			return err
-		}
-		_, err = impl.pipelineBuilder.DeleteCdPipeline(&pipeline, context.Background(), bean.FORCE_DELETE, false, 1, cdPipeline)
+		_, err = impl.pipelineBuilder.DeleteCdPipeline(&pipeline, context.Background(), bean.FORCE_DELETE, false, 1)
 		if err != nil {
 			impl.logger.Errorw("error in deleting cd pipeline", "err", err)
 			return err
