+++ pkg/pipeline/WorkflowUtils.go
@@ -5,12 +5,8 @@ import (
 	"github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
 	bean2 "github.com/devtron-labs/devtron/api/bean"
 	"github.com/devtron-labs/devtron/internal/sql/repository"
-	"github.com/devtron-labs/devtron/internal/sql/repository/pipelineConfig"
 	"github.com/devtron-labs/devtron/pkg/pipeline/bean"
-	repository2 "github.com/devtron-labs/devtron/pkg/pipeline/repository"
-	repository3 "github.com/devtron-labs/devtron/pkg/plugin/repository"
 	"github.com/devtron-labs/devtron/util"
-	"gopkg.in/yaml.v2"
 	v12 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"strconv"
@@ -22,23 +18,6 @@ type ConfigMapSecretDto struct {
 	Name     string
 	Data     map[string]string
 	OwnerRef v1.OwnerReference
-}
-
-type TaskYaml struct {
-	Version          string             `yaml:"version"`
-	CdPipelineConfig []CdPipelineConfig `yaml:"cdPipelineConf"`
-}
-type CdPipelineConfig struct {
-	BeforeTasks []*Task `yaml:"beforeStages"`
-	AfterTasks  []*Task `yaml:"afterStages"`
-}
-type Task struct {
-	Id             int    `json:"id"`
-	Index          int    `json:"index"`
-	Name           string `json:"name" yaml:"name"`
-	Script         string `json:"script" yaml:"script"`
-	OutputLocation string `json:"outputLocation" yaml:"outputLocation"` // file/dir
-	RunStatus      bool   `json:"-"`                                    // task run was attempted or not
 }
 
 func ExtractVolumesFromCmCs(configMaps []bean2.ConfigSecretMap, secrets []bean2.ConfigSecretMap) []v12.Volume {
@@ -283,88 +262,4 @@ func AddTemplatesForGlobalSecretsInWorkflowTemplate(globalCmCsConfigs []*bean.Gl
 	}
 
 	return nil
-}
-
-func StageYamlToPipelineStageAdapter(stageConfig string, stageType repository2.PipelineStageType, triggerType pipelineConfig.TriggerType) (*bean.PipelineStageDto, error) {
-	//sample stageConfig:= "version: 0.0.1\ncdPipelineConf:\n  - afterStages:\n      - name: test-1\n        script: |\n          date > test.report\n          echo 'hello'\n        outputLocation: ./test.report\n      - name: test-2\n        script: |\n          date > test2.report\n        outputLocation: ./test2.report"
-
-	pipelineStageDto := &bean.PipelineStageDto{}
-	var err error
-	taskYamlObject, err := ToTaskYaml([]byte(stageConfig))
-	if err != nil {
-		return nil, err
-	}
-	for _, task := range taskYamlObject.CdPipelineConfig {
-		if len(task.BeforeTasks) > 0 {
-			beforeStepIndex := 1
-			var beforeStepDtos []*bean.PipelineStageStepDto
-			for _, beforeTask := range task.BeforeTasks {
-				inlineStepDetail := &bean.InlineStepDetailDto{
-					ScriptType: repository3.SCRIPT_TYPE_SHELL,
-					Script:     beforeTask.Script,
-				}
-				//index really matters as the task order on the UI is decided by the index field
-				stepData := &bean.PipelineStageStepDto{
-					Id:                  0,
-					Name:                beforeTask.Name,
-					Description:         "",
-					Index:               beforeStepIndex,
-					StepType:            repository2.PIPELINE_STEP_TYPE_INLINE,
-					OutputDirectoryPath: []string{beforeTask.OutputLocation},
-					InlineStepDetail:    inlineStepDetail,
-					RefPluginStepDetail: nil,
-				}
-				beforeStepDtos = append(beforeStepDtos, stepData)
-				beforeStepIndex++
-			}
-			pipelineStageDto.Steps = beforeStepDtos
-			pipelineStageDto.Type = stageType
-			pipelineStageDto.Id = 0
-			if triggerType != pipelineConfig.TRIGGER_TYPE_AUTOMATIC && triggerType != pipelineConfig.TRIGGER_TYPE_MANUAL {
-				pipelineStageDto.TriggerType = pipelineConfig.TRIGGER_TYPE_MANUAL
-			} else {
-				pipelineStageDto.TriggerType = triggerType
-			}
-
-		}
-
-		if len(task.AfterTasks) > 0 {
-			afterStepIndex := 1
-			var afterStepDtos []*bean.PipelineStageStepDto
-			for _, afterTask := range task.AfterTasks {
-				inlineStepDetail := &bean.InlineStepDetailDto{
-					ScriptType: repository3.SCRIPT_TYPE_SHELL,
-					Script:     afterTask.Script,
-				}
-				stepData := &bean.PipelineStageStepDto{
-					Id:                  0,
-					Name:                afterTask.Name,
-					Description:         "",
-					Index:               afterStepIndex,
-					StepType:            repository2.PIPELINE_STEP_TYPE_INLINE,
-					OutputDirectoryPath: []string{afterTask.OutputLocation},
-					InlineStepDetail:    inlineStepDetail,
-					RefPluginStepDetail: nil,
-				}
-				afterStepDtos = append(afterStepDtos, stepData)
-				afterStepIndex++
-			}
-			pipelineStageDto.Steps = afterStepDtos
-			pipelineStageDto.Type = stageType
-			pipelineStageDto.Id = 0
-			if triggerType != pipelineConfig.TRIGGER_TYPE_AUTOMATIC && triggerType != pipelineConfig.TRIGGER_TYPE_MANUAL {
-				pipelineStageDto.TriggerType = pipelineConfig.TRIGGER_TYPE_MANUAL
-			} else {
-				pipelineStageDto.TriggerType = triggerType
-			}
-		}
-	}
-
-	return pipelineStageDto, nil
-}
-
-func ToTaskYaml(yamlFile []byte) (*TaskYaml, error) {
-	taskYaml := &TaskYaml{}
-	err := yaml.Unmarshal(yamlFile, taskYaml)
-	return taskYaml, err
 }
